+++ calendar-bundle/CHANGELOG.md
@@ -2,4 +2,5 @@
 
 ### DEV
 
+ * Add the teaser image as media:content in the RSS/Atom feeds (see contao/news-bundle#7).
  * Support sorting enclosures (see #16).